refactor(wayland): split state and input monitor modules
